Avoid colons in image filenames
Colons in filenames are anathema to the Windows kernel.

This commit produces PFS distribution archives which are incompatible with
those of v0.1.0. The new `pfs image push` fails with an old archive and the
old `pfs image push` fails with a new archive. That's ok because v0.1.0 was
only an alpha release.

Fixes https://github.com/pivotal-cf/pfs/issues/185